% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_longwave.R
\name{add_longwave}
\alias{add_longwave}
\title{Add longwave to ensemble forecast dataframe using Idso and Jackson (1969)}
\usage{
add_longwave(df)
}
\arguments{
\item{df}{}
}
\description{
Add longwave to ensemble forecast dataframe using Idso and Jackson (1969)
}
