% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_efi_standard.R
\name{convert_to_efi_standard}
\alias{convert_to_efi_standard}
\title{Convert units and names to CF and Ecological Forecasting Initiative standard}
\usage{
convert_to_efi_standard(df)
}
\arguments{
\item{df}{data frame output by get_ensemble_forecast}
}
\value{
data frame
}
\description{
Output units:
\itemize{
\item air_temperature: K
\item relative_humidity: proportion
\item surface_downwelling_longwave_flux_in_air: W m-2
\item surface_downwelling_shortwave_flux_in_air: W m-2
\item precipitation_flux: kg m-2 s-1
\item wind_speed: m s-1
\item air_pressure: Pa
\item cloud_cover: proportion
}
}
\examples{
get_ensemble_forecast(
latitude = 37.30,
longitude = -79.83,
forecast_days = 7,
past_days = 2,
model = "gfs_seamless",
variables = RopenMeteo::glm_variables(product = "ensemble_forecast",
                                      time_step = "hourly")) |>
  add_longwave() |>
  convert_to_efi_standard()

}
