% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmeteo.R
\name{open_meteo_build_url}
\alias{open_meteo_build_url}
\title{build url (internal)}
\usage{
open_meteo_build_url(
  type = "archive",
  latitude = NULL,
  longitude = NULL,
  start_date = NULL,
  end_date = NULL,
  variables = NULL,
  models = NULL
)
}
\description{
build url (internal)
}
