% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmeteo.R
\name{open_meteo_get_data}
\alias{open_meteo_get_data}
\title{get hourly (!!) data form openmeteoapi}
\usage{
open_meteo_get_data(
  type = "archive",
  latitude = 46.49,
  longitude = 11.34,
  start_date = today() - 50,
  end_date = today(),
  variables,
  models = NULL
)
}
\arguments{
\item{type}{chose "forecast" or "archive"}

\item{latitude}{latitude}

\item{longitude}{longitude}

\item{start_date}{a start date}

\item{end_date}{end date}

\item{variables}{a vector of variables}
}
\description{
get hourly (!!) data form openmeteoapi
}
