% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_url.R
\name{read_url}
\alias{read_url}
\title{Read open-meteo url to JSON}
\usage{
read_url(url, max_tries = 3)
}
\arguments{
\item{url}{API url}

\item{max_tries}{number tries before fails}
}
\value{
list
}
\description{
Read open-meteo url to JSON
}
