% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_glm_format.R
\name{write_glm_format}
\alias{write_glm_format}
\title{Write ensemble forecast dataframe to GLM formated csv files}
\usage{
write_glm_format(df, path)
}
\arguments{
\item{df}{data frame output by get_ensemble_forecast}

\item{path}{directory where csv files will be written}
}
\description{
Write ensemble forecast dataframe to GLM formated csv files
}
