% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_longwave.R
\name{add_longwave}
\alias{add_longwave}
\title{Add longwave to ensemble forecast dataframe using Idso and Jackson (1969)}
\usage{
add_longwave(df)
}
\arguments{
\item{df}{data frame output by get_ensemble_forecast}
}
\value{
data frame (in long format)
}
\description{
Add longwave to ensemble forecast dataframe using Idso and Jackson (1969)
}
