% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_longwave.R
\name{add_longwave}
\alias{add_longwave}
\title{Add longwave to ensemble forecast dataframe using Idso and Jackson (1969).  Requires cloud cover and temperature variables in input data frame.}
\usage{
add_longwave(df)
}
\arguments{
\item{df}{data frame output by \code{get_ensemble_forecast()}}
}
\value{
data frame (in long format)
}
\description{
Add longwave to ensemble forecast dataframe using Idso and Jackson (1969).  Requires cloud cover and temperature variables in input data frame.
}
\examples{
get_ensemble_forecast(
    latitude = 37.30,
    longitude = -79.83,
    forecast_days = 7,
    past_days = 2,
    model = "gfs_seamless",
    variables = c("cloud_cover", "temperature_2m")) |>
add_longwave()

}
