% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_efi_standard.R
\name{convert_to_efi_standard}
\alias{convert_to_efi_standard}
\title{Convert units and names to CF and EFI standard}
\usage{
convert_to_efi_standard(df)
}
\arguments{
\item{df}{data frame output by get_ensemble_forecast}
}
\value{
data frame
}
\description{
Convert units and names to CF and EFI standard
}
