% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_to_hourly.R
\name{daily_to_hourly}
\alias{daily_to_hourly}
\title{Convert daily climate projections to hourly time-step}
\usage{
daily_to_hourly(df, latitude, longitude)
}
\arguments{
\item{df}{data frame output by get_seasonal_forecast}

\item{latitude}{latitude degree north}

\item{longitude}{long longitude degree east or degree west}
}
\value{
data frame
}
\description{
Convert daily climate projections to hourly time-step
}
