% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_projections.R
\name{get_climate_projections}
\alias{get_climate_projections}
\title{Download point-level climate projections using open-meteo API}
\usage{
get_climate_projections(
  latitude,
  longitude,
  start_date,
  end_date,
  model = "EC_Earth3P_HR",
  variables = c("temperature_2m_mean")
)
}
\arguments{
\item{latitude}{latitude degree north}

\item{longitude}{long longitude degree east or degree west}

\item{start_date}{Number of days in the future for forecast (starts at current day)}

\item{end_date}{Number of days in the past to include in the data}

\item{variables}{vector of name of variable(s) https://open-meteo.com/en/docs/ensemble-api}
}
\value{
data frame
}
\description{
Download point-level climate projections using open-meteo API
}
