% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensemble_forecast.R
\name{get_ensemble_forecast}
\alias{get_ensemble_forecast}
\title{Download point-level ensemble weather forecasting using open-meteo API}
\usage{
get_ensemble_forecast(
  latitude,
  longitude,
  horizon,
  hist_days,
  model = "gfs_seamless",
  variables = c("relativehumidity_2m", "precipitation", "windspeed_10m", "cloudcover",
    "temperature_2m", "shortwave_radiation")
)
}
\arguments{
\item{variables}{}
}
\description{
Download point-level ensemble weather forecasting using open-meteo API
}
