% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_weather.R
\name{get_historical_weather}
\alias{get_historical_weather}
\title{Download point-level historical weather (ERA5) using open-meteo API}
\usage{
get_historical_weather(
  latitude,
  longitude,
  site_id = NULL,
  start_date,
  end_date,
  variables = c("relativehumidity_2m", "precipitation", "windspeed_10m", "cloudcover",
    "temperature_2m", "shortwave_radiation")
)
}
\arguments{
\item{latitude}{latitude degree north}

\item{longitude}{long longitude degree east or degree west}

\item{site_id}{name of site location (optional, default = NULL)}

\item{start_date}{earliest date requested. Must be on or after 1950-01-01}

\item{end_date}{latest date requested}

\item{variables}{vector of name of variable(s) https://open-meteo.com/en/docs/ensemble-api}
}
\value{
data frame
}
\description{
Download point-level historical weather (ERA5) using open-meteo API
}
