% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_variables.R
\name{glm_variables}
\alias{glm_variables}
\title{Get set of variables required for the GLM model}
\usage{
glm_variables(product, time_step)
}
\arguments{
\item{product}{api type: climate, forecast, ensemble_forecast, historical, seasonal_forecast}

\item{time_step}{model and time-step: hourly, 6hour, daily}
}
\value{
character vector
}
\description{
Get set of variables required for the GLM model
}
