% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/six_hourly_to_hourly.R
\name{six_hourly_to_hourly}
\alias{six_hourly_to_hourly}
\title{Convert 6 hour seasonal forecast to hourly time-step}
\usage{
six_hourly_to_hourly(df, latitude, longitude, use_solar_geom = TRUE)
}
\arguments{
\item{df}{data frame output by get_seasonal_forecast}

\item{latitude}{latitude degree north}

\item{longitude}{long longitude degree east or degree west}

\item{use_solar_geom}{use solar geometry to determine hourly variables}
}
\value{
data frame
}
\description{
Convert 6 hour seasonal forecast to hourly time-step
}
